arm64: fpsimd: dovetail: enable alternate scheduling
Signed-off-by: Philippe Gerum <rpm@xenomai.org>

arm64: fpsimd: dovetail: Allow preemption of in-band fpsimd kernel sections

Preemption of in-band in-kernel fpsimd sections was not allowed /
properly implemented in Dovetail up to now.

That piled up again as some crypto code was optimized for arm64 in
Linux 6.14. A possible call stack that makes clear that the FPU was
corrupted follows. The inband stage was using crypto/FPU code inside
the kernel and got interrupted by OOB.

	[  331.703695]  task_fpsimd_load+0xd0/0x214 (P)
	[  331.703713]  fpsimd_restore_current_state+0x9c/0xb8
	[  331.703732]  fpsimd_restore_current_oob+0x1c/0x2c
	[  331.703751]  dovetail_context_switch+0x90/0x180
	[  331.703769]  pipeline_switch_to+0x10/0x1c
	[  331.703792]  ___xnsched_run+0x1bc/0x29c
	[  331.703811]  run_oob_call+0x98/0x17c
	[  331.703830]  handle_irq_pipelined_finish+0x1e0/0x1ec
	[  331.703848]  handle_irq_pipelined+0x58/0x6c
	[  331.703865]  call_on_irq_stack+0x24/0x4c
	[  331.703884]  do_interrupt_handler+0x150/0x178
	[  331.703902]  el1_interrupt+0x44/0x11c
	[  331.703921]  el1h_64_irq_handler+0x18/0x24
	[  331.703941]  el1h_64_irq+0x6c/0x70                <- oob
	[  331.703957]  crc32c_le_arm64_4way+0x4c/0xc0 (P)   <- inband
	[  331.703978]  ext4_block_bitmap_csum_set+0x38/0x60
	[  331.704002]  ext4_mb_mark_context+0x1b0/0x398
	[  331.704025]  ext4_mb_mark_diskspace_used+0xd8/0x1f8

To fix that we have to hook into the task switch as we do on x86 and
invoke the FPU (re)store logic.

Signed-off-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Acked-by: Philippe Gerum <rpm@xenomai.org>

arm64: fpsimd: dovetail: Align with x86 implementation

The FPU preemption marker is only accessed from the current task, so
we can avoid atomic operations. Moving the marker from the low level
flags to the local (synchronous) flags to make that clear.

Gives us some performance and aligns the arm64 implementation with x86.

Signed-off-by: default avatarFlorian Bezdeka <florian.bezdeka@siemens.com>

arm64: fpsimd: dovetail: drop redundant hard irqs disabling

fpsimd_save_and_flush_cpu_state() itself does take care of this.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>