evl/mutex: sanitize dropping of owned mutexes at thread exit
We should never rely on fundles to determine which mutexes should be
dropped by an exiting thread: this data is shared with userland,
therefore it is unsafe by nature.

Since all mutexes a thread owns is duly tracked by the core, trust the
tracking list exclusively to figure out which mutexes need to be
released, there is no rationale to revalidate this information via
fundle probing.

At this chance, use the common thread->lock to serialize access to
this list, there is no point in throwing yet another lock dedicated to
tracking, since we have to hold it anyway for the surrounding
operations (locking/unlocking), or nobody racing with us (i.e. at
thread exit). This also fixes an ABBA issue which existed between
wchan->lock and thread->tracking_lock. Now we only need to hold
wchan->lock then thread->lock nested, which is deadlock-safe.

Signed-off-by: Philippe Gerum <rpm@xenomai.org>