evl/net: allow for eBPF-based filtering on early RX path
We can use eBPF filtering in order to leave the decision about oob vs
inband packet delivery to userland. Provide an ioctl() interface to do
so on a per-device basis.

An eBPF filter may be attached to any netdevice which is enabled for
oob diversion. If present, this program is in charge of determining
which stack is eventually going to handle every ingress packet passed
to netif_oob_deliver() by the inband stack. The eBPF program belongs
to the "socket" class, since we already have a skb at this point.

In absence of filter, the default VLAN-based rules apply, i.e. an
ingress packet should be handled by the oob stack if it is VLAN
tagged, and its VLAN id matches one of the ids defined as oob traffic
discriminators.