evl/net: arp: simplify pseudo-entry handling
We currently maintain a full-fledged, dynamically allowed, refcounted
and semi-permanent ARP entry for the loopback address. This is 100%
overkill.

We only need the user to provide a pseudo-entry descriptor upon cache
lookup, which we would fill on the fly if the loopback address is
matched since the resolved output in that case is constant.

Simplify the implementation accordingly, fixing up its sole user at
the moment (UDP).

Signed-off-by: Philippe Gerum <philippe.gerum@exail.com>